ARM: shmobile: bockw: Do not set command line in defconfig
As a set towards moving shmobile towards multi-platform
the kernel command line should be set in the DTB rather than
in the kernel config.

bockw already has the command line present in its DTS file
but it was being overrirden by the kernel config.

A side effect of this change is that "rw" is now present
on the command line which allows a boot using nfs root to
succeed in the case where userspace expects to be able
to write to the root filesystem.

Signed-off-by: Simon Horman <horms@verge.net.au>
(cherry picked from commit 851f0cda469fdba6371ab2e268462aeb8393f08e)
(Queued by ARM-SoC for v3.14 but not yet in Linus's tree)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>