v4l: vsp1: Fix LUT support
This patch corrects a error in the backport of
c655ad627a647e51 ("v4l: vsp1: Add LUT support").

In mainline the vsp1 driver directory is drivers/media/platform/vsp1/
in keeping with the location of other media platform drivers.
However, in the ltsi-3.4.25 backports the driver directory is
drivers/media/video/vsp1/, which is in keeping with the location
of other media platform drivers in 3.4.

Accordingly, the backports update the driver, placing its files
in the legacy directory. However vsp1_lut.h was placed in its
mainline rather than the legacy location when backported.
This patch resolves that problem.

Signed-off-by: Simon Horman <horms+renesas@verge.net.au>