i2c: rcar: use devm_clk_get to ensure clock is properly ref-counted
The current i2c-rcar driver does clk_get() without a corresponding
clk_put(). Add the clk to the driver private data and then get it
with the devm functions so that it is released when the driver is
unbound.

Note, we do not call clk_prepare_enable() at this point due to the
very possible magic that is being done by the pm_runtime system
underneath the driver.

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
(cherry picked from commit bc8120f17ae87da0850b4e6a806ad88ffd01ca64)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>