local: serial: sh-sci: Add support DMA of HSCIF
This supports DMA of HSCIF with handling of timeout for receiving DMA data.
Timeout for receiving DMA data of HSCIF use timer interrupt. And timer
interrupt is set in DMA receive callback function. Since the DMA receive
callback function is not called until the data accumulated in buffer,
data will not be sent immediately.
In order to address this issue for HSCIF, this patch increase to receive
buffer. Then, first buffer is used to call the DMA receive callback function
immediately.

Signed-off-by: Nobuhiro Iwamatsu <nobuhiro.iwamatsu.yj@renesas.com>