dma: sh: Change error log in shdma_tx_submit
When AudioDMAC-pp is used, the error message is displayed in spite of
no error. This message interferes with the output of an audio head.
In order to avoid it, the output level of the error log was changed
into "dev_dbg" from "dev_err".

Signed-off-by: j-watana <j-watana@RI116989.adwin.renesas.com>
(cherry picked from commit 529cd85ffaefaa928f289fb20e5ee3ca1b81ab78)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>