media: soc_camera: rcar_vin: Add capture width check for NV16 format
At the time of NV16 capture format, the user has to specify the
capture output width of the multiple of 32 for H/W specification.
At the time of using NV16 format by ioctl of VIDIOC_S_FMT,
this patch adds align check and the error handling to forbid
specification of the capture output width which is not a multiple of 32.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
(cherry picked from commit 2823a5b271c1bffbb4ba751efed40a772db50a8f)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>