drm: rcar-du: Fix LVDS output data gap workaround
There was an miscalculation of the wait time by loop. Since the
number of CPU instruction of loop is three, the number of loop
is divided by 3. Since CPU has 3way superscaler, the number of loop
is multiplied by 3. Moreover, the wait time is corrected not the middle
time but the minimum time for the CPU load processing occurring
occasionally.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
(cherry picked from commit 85b1b9c4681633874dccf171e2cf6c6a43891ddd)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>