ARM: shmobile: armadillo800eva dts: Add chosen/stdout-path
Add a stdout-path property so that automatic console selection works
in the absence of a "console=" parameter on the kernel command line.
In addition, this will allow the PM domain code to find the PM domain
for the console device.

Note that we have to keep the "console=ttySC1" parameter in
chosen/bootargs, else we only get console messages on tty0, and because
this DTS is shared between legacy and multi-platform.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
(cherry picked from commit 2c32622c3f3f9c7e0d3480c1fc0a31a95e04dd91)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>