ARM: shmobile: lager: Fix register area of ADV7511 in device tree
This patch moves registration of the ADV7511.
If the ADV7511 is registered in the i2c group,
I2C driver outputs log as follows.

  - i2c i2c-2: of_i2c: invalid reg on /i2c@e6530000/adv7511

Since the ADV7511 is registered as slave device in the
DRM driver, it should not register in the group of i2c.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
(cherry picked from commit 9064f4e6e4074a15e6510e8a56ec9e3c3ecfd91f)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>