drm: rcar-du: Fix dot clock selection method
The value of mode->clock may turn into an unexpected value
when EDID is not able to read or EDID is injustice.
In that case, the value of div_ex may be set to 0 and may
do division by 0. This problem is resolved with this patch.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
(cherry picked from commit dea9547c1c2d27ac6d8597ca36b57e593778099b)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>